+++ common-custom-user-data-maven-extension/src/main/java/com/gradle/CustomBuildScanEnhancements.java
@@ -64,7 +64,6 @@ final class CustomBuildScanEnhancements {
             Optional<String> teamCityConfigFile = projectProperty(mavenSession, "teamcity.configuration.properties.file");
             Optional<String> buildNumber = projectProperty(mavenSession, "build.number");
             Optional<String> buildTypeId = projectProperty(mavenSession, "teamcity.buildType.id");
-            Optional<String> agentName = projectProperty(mavenSession, "agent.name");
             if (teamCityConfigFile.isPresent()
                     && buildNumber.isPresent()
                     && buildTypeId.isPresent()) {
@@ -77,7 +76,9 @@ final class CustomBuildScanEnhancements {
             }
             buildNumber.ifPresent(value ->
                     buildScan.value("CI build number", value));
-            agentName.ifPresent(value ->
+            buildTypeId.ifPresent(value ->
+                    addCustomValueAndSearchLink(buildScan, "CI build config", value));
+            projectProperty(mavenSession, "agent.name").ifPresent(value ->
                     addCustomValueAndSearchLink(buildScan, "CI agent", value));
         }